requirements.txt
altgraph==0.17.4
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
importlib_metadata==7.0.2
macholib==1.16.3
packaging==24.0
paramiko==3.4.0
pycparser==2.21
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyNaCl==1.5.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PyQt5-stubs==5.15.6.0
pyspnego==0.10.2
pywinrm==0.4.3
requests==2.31.0
requests-ntlm==1.2.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
urllib3==2.2.1
xmltodict==0.13.0
zipp==3.18.0
